Merged from develop - several Bugfixes

+++ classes/Course/class.hubCourse.php
@@ -17,6 +17,10 @@ class hubCourse extends hubRepositoryObject {
 	 */
 	public $ilias_object;
 
+    /**
+     * @var int
+     */
+    public static $id_type = self::ILIAS_ID_TYPE_REF_ID;
 
 	/**
 	 * @return string
@@ -61,12 +65,14 @@ class hubCourse extends hubRepositoryObject {
 					hubCounter::incrementUpdated($hubCourse->getSrHubOriginId());
 					break;
 				case hubSyncHistory::STATUS_DELETED:
+                    global $tree;
+                    $path = $tree->getPathId($history->getIliasId());
 					if (! hubSyncCron::getDryRun()) {
 						$hubCourse->deleteCourse();
 						$hubOriginObj->afterObjectDeletion($hubCourse);
 					}
 					hubCounter::incrementDeleted($hubCourse->getSrHubOriginId());
-					hubOriginNotification::addMessage($hubCourse->getSrHubOriginId(), $full_title, 'Courses deleted:');
+					hubOriginNotification::addMessage($hubCourse->getSrHubOriginId(), $full_title . ' :: ' . implode('/', $path), 'Courses deleted with ref_id path:');
 					break;
 				case hubSyncHistory::STATUS_ALREADY_DELETED:
 					hubCounter::incrementIgnored($hubCourse->getSrHubOriginId());
@@ -83,7 +89,9 @@ class hubCourse extends hubRepositoryObject {
 
 			$history->updatePickupDate();
 			$hubOrigin::afterObjectModification($hubCourse);
-			$hubOriginObj->afterObjectInit($hubCourse);
+            if (! hubSyncCron::getDryRun()) {
+                $hubOriginObj->afterObjectInit($hubCourse);
+            }
 
 			hubDurationLogger2::getInstance($id)->pause();
 		}
@@ -97,16 +105,20 @@ class hubCourse extends hubRepositoryObject {
 		$this->ilias_object->setTitle($this->getTitlePrefix() . $this->getTitle() . $this->getTitleExtension());
 		$this->ilias_object->setDescription($this->getDescription());
 		$this->ilias_object->setImportId($this->returnImportId());
-		$this->ilias_object->setImportantInformation($this->getIm());
-		if ($this->props()->get(hubCourseFields::F_ACTIVATE)) {
-			$this->ilias_object->setActivationType(IL_CRS_ACTIVATION_UNLIMITED);
-		}
+		$this->ilias_object->setImportantInformation($this->getImportantInformation());
 		$this->updateAdditionalFields();
 		$this->ilias_object->create();
 		$this->ilias_object->createReference();
 		$node = $this->getDependecesNode();
 		$this->ilias_object->putInTree($node);
 		$this->ilias_object->setPermissions($node);
+        $this->ilias_object->setSubscriptionLimitationType($this->getSubLimitationType());
+        $this->ilias_object->updateSettings();
+		if ($this->props()->get(hubCourseFields::F_ACTIVATE)) {
+			$this->ilias_object->setOfflineStatus(false);
+			$this->ilias_object->setActivationType(IL_CRS_ACTIVATION_UNLIMITED);
+			$this->ilias_object->update();
+		}
 		if ($this->props()->get(hubCourseFields::F_CREATE_ICON)) {
 			$this->updateIcon($this->ilias_object);
 			$this->ilias_object->update();
@@ -117,7 +129,13 @@ class hubCourse extends hubRepositoryObject {
 			global $ilSetting;
 			$mail = new ilMimeMail();
 			$mail->autoCheck(false);
-			$mail->From($ilSetting->get('admin_email'));
+            if($this->props()->get(hubCourseFields::F_NOT_FROM)){
+                $mail->From($this->props()->get(hubCourseFields::F_NOT_FROM));
+            }
+            else{
+                $mail->From($ilSetting->get('admin_email'));
+            }
+
 			$mail->To($this->getNotificationEmail());
 			$body = hubCourseFields::getReplacedText($this);
 			$mail->Subject($this->props()->get(hubCourseFields::F_NOT_SUBJECT));
@@ -138,14 +156,19 @@ class hubCourse extends hubRepositoryObject {
 		if ($this->props()->get(hubCourseFields::F_UPDATE_TITLE)) {
 			$this->initObject();
 			$this->ilias_object->setTitle($this->getTitlePrefix() . $this->getTitle() . $this->getTitleExtension());
-			$this->ilias_object->setDescription($this->getDescription());
-			$update = true;
+            $update = true;
 		}
 		if ($this->props()->get(hubCourseFields::F_UPDATE_DESCRIPTION)) {
 			$this->initObject();
 			$this->ilias_object->setDescription($this->getDescription());
 			$update = true;
 		}
+        if ($this->props()->get(hubCourseFields::F_UPDATE_RESPONSIBLE)) {
+            $this->initObject();
+            $this->ilias_object->setContactResponsibility($this->getResponsible());
+            $this->ilias_object->setContactEmail($this->getResponsibleEmail());
+            $update = true;
+        }
 		if ($this->props()->get(hubCourseFields::F_UPDATE_ICON)) {
 			$this->initObject();
 			$this->updateIcon($this->ilias_object);
@@ -153,11 +176,12 @@ class hubCourse extends hubRepositoryObject {
 		}
 		if ($this->props()->get(hubCourseFields::F_REACTIVATE)) {
 			$this->initObject();
+			$this->ilias_object->setOfflineStatus(false);
 			$this->ilias_object->setActivationType(IL_CRS_ACTIVATION_UNLIMITED);
 			$update = true;
 		}
 		if ($update) {
-			$this->updateAdditionalFields();
+            $this->ilias_object->setOwner($this->getOwner());
 			$this->ilias_object->update();
 		}
 
@@ -174,7 +198,6 @@ class hubCourse extends hubRepositoryObject {
 		}
 		$this->ilias_object->setContactResponsibility($this->getResponsible());
 		$this->ilias_object->setContactEmail($this->getResponsibleEmail());
-		$this->ilias_object->setSubscriptionLimitationType($this->getSubLimitationType());
 		$this->ilias_object->setOwner($this->getOwner());
 	}
 
@@ -196,8 +219,9 @@ class hubCourse extends hubRepositoryObject {
 					$this->ilias_object->update();
 					break;
 				case self::DELETE_MODE_DELETE:
-					$this->ilias_object->delete();
-					$hist->setIliasId(NULL);
+					if($this->ilias_object) {
+						$this->ilias_object->delete();
+					}
 					break;
 				case self::DELETE_MODE_TRASH:
 					global $tree;
